restructured generators crate
