+++ src/dnl-generators/src/scenarios/battle.rs
