+++ src/dnl-generators/src/scenarios/mod.rs
@@ -1,0 +1,3 @@
+pub mod battle;
+pub mod rest;
+pub mod shop;
