+++ src/dnl-generators/src/scenarios/rest.rs
