+++ src/dnl-generators/src/scenarios/shop.rs
