+++ src/dnl-generators/src/tools/mod.rs
@@ -1,0 +1,1 @@
+pub mod random_weapon;
